chore: 重構專案資料夾 projrcts/mnist/
