--- flows/flow-mnist/config/dataset.yml
@@ -1,5 +1,0 @@
-# 訓練資料相關設定
-train_data_path: './MNIST/train'  # 訓練資料相對於flows_mnist的相對路徑
-dvc_remote: 's3://dvc/mnist/dvc_remote'  # DVC remote
-dvc_s3_endpoint_url: 'http://minio:9000'
-dvc_endpoint_for_client: 'minio:9000'
