--- flows/flows_mnist_gpu_0/requirements.txt
@@ -1,8 +1,0 @@
-numpy~=1.23.5
-python-dotenv~=0.21.0
-torchvision
-torch==2.0.*
-gdown
-dvc
-dvc_s3
-pyyaml
