--- flows/flows_mnist_gpu_1/.prefectignore
@@ -1,8 +1,0 @@
-.dvc/
-MNIST/
-MNIST.zip
-__pycache__/
-.DS_Store/
-
-expand_train_data.py
-demo_create_first_version_dvc.sh
