--- mlops-sys/flow_agent/mnist-cpu_mnist_local_cpu/.env
@@ -1,18 +1,0 @@
-# MinIO
-MINIO_ROOT_USER='admin'
-MINIO_ROOT_PASSWORD='adminsecretkey'
-MLFLOW_BUCKET_NAME='mlflow'
-MLFLOW_S3_ENDPOINT_URL='http://minio:9000'
-
-# postgres
-MLFLOW_TRACKING_URI='http://mlflow_server:5050'
-
-# prefect
-PREFECT_API_URL='http://prefect_server:4200/api'
-
-
-### self settings ###
-# The agent will run the shedules with this setting
-POOL_NAME='mnist-cpu'
-QUEUE_NAME='mnist'
-
