+++ mlops-sys/flow_agent/mnist-cpu_mnist_singal_cpu/.env
@@ -1,0 +1,17 @@
+# MinIO
+MINIO_ROOT_USER='admin'
+MINIO_ROOT_PASSWORD='adminsecretkey'
+MLFLOW_BUCKET_NAME='mlflow'
+MLFLOW_S3_ENDPOINT_URL='http://minio:9000'
+
+# Prefect
+PREFECT_API_URL='http://prefect_server:4200/api'
+
+# MLflow
+AWS_ACCESS_KEY_ID="${MINIO_ROOT_USER}"
+AWS_SECRET_ACCESS_KEY="${MINIO_ROOT_PASSWORD}"
+MLFLOW_TRACKING_URI='http://mlflow_server:5050'
+
+### Prefect Agent
+POOL_NAME='mnist-cpu'
+QUEUE_NAME='mnist-cpu'
