+++ mlops-sys/flow_agent/mnist-cpu_mnist_singal_cpu/docker-compose.yml
