+++ mlops-sys/flow_agent/mnist-gpu_mnist_single_gpu/.env
@@ -1,0 +1,20 @@
+# 設定 - flow agent 將依照此設定運行排程
+SERVER_IP='YOUR_SERVER_IP'      # Server IP
+
+# MinIO
+MINIO_ROOT_USER='admin'
+MINIO_ROOT_PASSWORD='adminsecretkey'
+MLFLOW_BUCKET_NAME='mlflow'
+MLFLOW_S3_ENDPOINT_URL='http://minio:9000'
+
+# Prefect
+PREFECT_API_URL='http://prefect_server:4200/api'
+
+# MLflow
+AWS_ACCESS_KEY_ID="${MINIO_ROOT_USER}"
+AWS_SECRET_ACCESS_KEY="${MINIO_ROOT_PASSWORD}"
+MLFLOW_TRACKING_URI='http://mlflow_server:5050'
+
+### Prefect Agent
+POOL_NAME='mnist-gpu'
+QUEUE_NAME='mnist-gpu'
