+++ mlops-sys/flow_agent/mnist-gpu_mnist_single_gpu/docker-compose.yml
@@ -1,0 +1,33 @@
+version: '3.8'
+
+name: flow_agent_${POOL_NAME}_${QUEUE_NAME}
+
+services:
+
+  flow_agent:
+    build: .
+    image: flow_agent_${POOL_NAME}_${QUEUE_NAME}_gpu
+    container_name: flow_agent_${POOL_NAME}_${QUEUE_NAME}_gpu
+    command: prefect agent start --pool ${POOL_NAME} --work-queue ${QUEUE_NAME}
+    environment:
+      # for mlflow
+      - AWS_ACCESS_KEY_ID=${MINIO_ROOT_USER}
+      - AWS_SECRET_ACCESS_KEY=${MINIO_ROOT_PASSWORD}
+      - MLFLOW_S3_ENDPOINT_URL=${MLFLOW_S3_ENDPOINT_URL}
+      - MLFLOW_BUCKET_NAME=${MLFLOW_BUCKET_NAME}
+      - MLFLOW_TRACKING_URI=${MLFLOW_TRACKING_URI}
+      # for prefect
+      - PREFECT_API_URL=${PREFECT_API_URL}
+    networks:
+      - mlops
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              device_ids: ['0']
+              capabilities: [gpu]
+
+networks:
+  mlops:
+    external: true
