+++ mlops-sys/flow_agent/mnist-gpu_mnist_single_gpu/requirements.txt
@@ -1,0 +1,12 @@
+# flow.py 會用到的相關套件 (除 mlflow==2.5.0 和 prefect==2.10.9 以外)
+dill
+numpy==1.26.3
+torchvision==0.15.2
+torch==2.0.1
+gdown==4.7.1
+dvc==3.33.2
+dvc_s3==2.23.0
+pyyaml==6.0.1
+torchinfo==1.8.0
+python-dotenv==1.0.0
+minio==7.2.2
