+++ mlops-sys/flow_agent/mnist-gpu_mnist_single_gpu/requirements_sys.txt
@@ -1,0 +1,14 @@
+# S3
+s3fs
+pyarrow
+fastparquet
+python-dotenv
+psycopg2-binary
+boto3
+
+# MLflow
+mlflow==2.5.0
+
+# DVC
+dvc
+dvc_s3
