+++ mlops-sys/flow_scheduler/.env.local
@@ -2,7 +2,7 @@
 # FLOW_DIR = '../../flows/example_flow' # project directory of your flow.py
 
 # MLflow
-MLFLOW_TRACKING_URI='postgresql://postgres:postgres@db:5432/mlrun'
+MLFLOW_TRACKING_URI='http://mlflow_server:5050'
 
 # Prefect
 PREFECT_API_URL='http://prefect_server:4200/api'
