+++ projects/mnist/dev/data/MNIST.dvc
@@ -1,0 +1,6 @@
+outs:
+- md5: 39d66d274c820c594ced12fd58323d07.dir
+  size: 39247104
+  nfiles: 70000
+  hash: md5
+  path: MNIST
