+++ projects/mnist/flow/config/flow.yml
@@ -1,0 +1,27 @@
+# 部署至Prefect伺服器的相關設定
+RFS_NAME: 'minio' # 值需與server/.env裡面的RFS_NAME環境變數相同
+
+deployment:
+  flow_py_name: flow.py  # flow.py的檔名
+
+  # 以下的flow_name與deploy_name會顯示在Prefect UI，為了方便管理，建議每個專案都有不同的flow_name與deploy_name名稱
+  flow_name: main  # 建議設定為專案名稱，與flow.py裡面以「@flow」裝飾的主函式名稱相同
+  deploy_name: model_training-gpu  # 目的，如daliy_model、data_update
+
+  pool_name: mnist-gpu  # 要調用的運算資源
+  queue_name: mnist-gpu  # 專案名稱
+
+# 排程相關設定
+scheduler:
+  # 排程方式設定
+  method: 'rrule'  # 排程的時間設定，可選擇'interval'、'cron'或'rrule'
+  timezone: 'Asia/Taipei'  # 使用的時區，在臺灣建議設定為'Asia/Taipei'
+  # 排程設定
+  interval: 3600  # 執行的時間間隔，單位為秒
+  cron: '1 * * * *'  # cron的設定值
+  rrule: 'FREQ=WEEKLY;BYDAY=MO,WE,FR;UNTIL=20240731T160000Z'  # rrule的設定值
+  # cron與rrule的撰寫格式教學：https://docs.prefect.io/2.10.21/concepts/schedules/
+
+# 實驗相關設定
+experiment_name: 'MNIST'  # 顯示在Prefect UI中的實驗名稱
+developer_name: 'AIF'
