+++ projects/mnist/img/MLflow_params.png
